# Optimize callcache invalidation for refinements