Fixes [Bug [#21201](https://bugs.ruby-lang.org/issues/21201)]

This change addresses a performance regression where defining methods inside `refine` blocks caused severe slowdowns. The issue was due to `rb_clear_all_refinement_method_cache()` triggering a full object space scan via `rb_objspace_each_objects` to find and invalidate affected callcaches, which is very inefficient.

To fix this, I introduce `vm->cc_refinement_table` to track callcaches related to refinements. This allows us to invalidate only the necessary callcaches without scanning the entire heap, resulting in significant performance improvement.

## benchmark

```
require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"

  gem "benchmark-ips"
end

mod = Module.new
klass = Class.new

Benchmark.ips do |x|
  x.report(RUBY_VERSION) do
    mod.send(:refine, klass) do
      def call_1 = nil
      def call_2 = nil
      def call_3 = nil
    end
  end

  x.save! "/tmp/performance_regression_refine.bench"
  x.compare!
end
```

```
ruby 3.5.0dev (2025-06-05T14:33:43Z bugs/21201 e7833e82f4) +PRISM [arm64-darwin24]
Warming up --------------------------------------
         3.5.0_patch    50.158k i/100ms
Calculating -------------------------------------
         3.5.0_patch    520.371k (± 7.2%) i/s    (1.92 μs/i) -      2.608M in   5.038852s

Comparison:
               3.2.8:  1597482.0 i/s
         3.5.0_patch:   520371.1 i/s - 3.07x  slower (this PR)
               3.3.8:      213.0 i/s - 7499.45x  slower
               3.4.4:      138.2 i/s - 11561.20x  slower
3.5.0_master_0e0008da0f19d:      137.8 i/s - 11593.43x  slower
```