@ko1 2025-04-07
I'm not sure about compaction, but is it safe for `GC.compact`?
cc/ @tenderlove