@byroot 2025-04-07
Nevermind, It's already handled because it's declared as a weak table.