@tenderlove 2025-04-07
> I'm not sure about compaction, but is it safe for `GC.compact`?

The reference updating code looks correct to me